Move all non-table-related stuff from VolumesTable to VolumesForm
